Migration finder skips modules starting with underscore
The new migration command always uses a timestamp at the start of the
migration file name, while custom files can break that convention
it is unlikely anyone will use _ as a starting character for migrations.

This largely is needed because otherwise the command thinks `__init__.py`
is a migration file, and because of sorting, it ends up last.

Fixes: #874